Remove fastAnalogRead
Refactored code to remove dependence on S3-FastAnalogRead. I don't
think that code was necessary with the overhauled ADC API in
ESP-IDF 5 (available for Arduino in ESP-Arduino 3.x). The code
now uses the continuous ADC API, which is over 2x faster than
single-shot using readAnalog. It reads 288 samples in under 5ms.

Now the main performance limitation is the clock. The C12880 supports
being clocked at up to 5MHz, but we are maxing out at less than
0.5MHz due to the use of delayMicros, which has a theoretical minimum
delay of 1us, but in practice is about 1.5us. Options for removing
this limit include:
* use a peripheral that can generate the pulses, like the RMT
  (http://esp32.io/viewtopic.php?t=37725)
* dedicated GPIO
  (https://docs.espressif.com/projects/esp-idf/en/latest/esp32s3/api-reference/peripherals/dedic_gpio.html)
* General-purpose timer (GPTimer). The Arduino API:
  https://docs.espressif.com/projects/arduino-esp32/en/latest/api/timer.html

Of these, the first is likely to be the most performant and should
be able to easily hit 5MHz with minimal CPU cycles. But the GPTimer
is probably the easilest to implement.