Adds nb & data for @benkrasnow's MultiSpect module
@benkrasnow was kind enough to lend me his MultiSpectLED module
(https://github.com/benkrasnow/MultiSpectLED) to measure the LED
spectra. A notebook to make the measurements and plot the spectra
is included, as well as the measurements that I made.

I learned a couple of things:
1. The C12880 spectrometer wavelength calibration seems accurate;
the measured peak values are close to the datasheet spec peaks.
The fact that the small differences go in both directions suggests
that the spectrometer readings are not biased one way or the other
(the most likely source of error.)
2. I did >= 30 measurements for each LED and the readings are very
consistent in terms of the relative spectra (ie., peak locations
align). However, there is work to do on the integration time
calibration; while the peak locations are aligned, the measured
voltages at the peak varied from one measurement to another.